backend/requirements.txt
fastapi
uvicorn
pydantic
openai>=1.0.0
python-dotenv
reportlab